+++ src/lib.rs
@@ -1,21 +1,20 @@
+use events::{Event, EventHandler, EventHandlers, HandlerFn};
 use hidapi::{DeviceInfo, HidApi};
-#[cfg(feature = "events")]
+
 use rayon::prelude::*;
 use std::{
     collections::HashMap,
     env::consts::OS,
     ops::BitOr,
     process::exit,
-    sync::{Arc, Mutex, RwLock},
+    sync::{atomic::AtomicBool, Arc, Mutex, RwLock},
     thread::{self, sleep},
     time::Duration,
-    vec,
 };
 
-type HandlerFn = fn(g29: &mut G29);
-
-pub mod state;
-
+pub mod events;
+// pub mod state;
+mod state;
 // The size of the data frame that the G29 sends
 const FRAME_SIZE: usize = 12;
 
@@ -40,80 +39,6 @@ pub enum DpadPosition {
 /// Event
 /// Events that can be triggered by the G29
 ///
-#[cfg(feature = "events")]
-#[derive(Debug, PartialEq, Copy, Clone, Eq, Hash)]
-pub enum Event {
-    /// Steering wheel is turned
-    Steering,
-    /// Steering wheel is turned finely
-    SteeringFine,
-    /// Throttle changed
-    Throttle,
-    /// Brake changed
-    Brake,
-    /// Clutch changed
-    Clutch,
-    DpadUpPressed,
-    DpadUpReleased,
-    DpadTopRightPressed,
-    DpadTopRightReleased,
-    DpadRightPressed,
-    DpadRightReleased,
-    DpadBottomRightPressed,
-    DpadBottomRightReleased,
-    DpadBottomPressed,
-    DpadBottomReleased,
-    DpadBottomLeftPressed,
-    DpadBottomLeftReleased,
-    DpadLeftPressed,
-    DpadLeftReleased,
-    DpadTopLeftPressed,
-    DpadTopLeftReleased,
-    XButtonPressed,
-    XButtonReleased,
-    SquareButtonPressed,
-    SquareButtonReleased,
-    CircleButtonPressed,
-    CircleButtonReleased,
-    TriangleButtonPressed,
-    TriangleButtonReleased,
-    RightShifterPressed,
-    RightShifterReleased,
-    LeftShifterPressed,
-    LeftShifterReleased,
-    R2ButtonPressed,
-    R2ButtonReleased,
-    L2ButtonPressed,
-    L2ButtonReleased,
-    ShareButtonPressed,
-    ShareButtonReleased,
-    OptionButtonPressed,
-    OptionButtonReleased,
-    R3ButtonPressed,
-    R3ButtonReleased,
-    L3ButtonPressed,
-    L3ButtonReleased,
-    PlusButtonPressed,
-    PlusButtonReleased,
-    MinusButtonPressed,
-    MinusButtonReleased,
-    // Spinner is rotating right
-    SpinnerRight,
-    // Spinner is rotating left
-    SpinnerLeft,
-    SpinnerButtonPressed,
-    SpinnerButtonReleased,
-    PlaystationButtonPressed,
-    PlaystationButtonReleased,
-    /// Shifter X axis changed
-    ShifterX,
-    /// Shifter Y axis changed
-    ShifterY,
-    ShifterPressed,
-    ShifterReleased,
-    /// Gear selector changed
-    GearChanged,
-}
 
 type Frame = [u8; FRAME_SIZE];
 
@@ -179,47 +104,7 @@ impl BitOr for Led {
     }
 }
 
-#[cfg(feature = "events")]
-#[derive(Debug, Copy, Clone)]
-pub struct EventHandler {
-    pub id: usize,
-    pub event: Event,
-    handler: HandlerFn,
-}
-
-#[cfg(feature = "events")]
-#[derive(Debug)]
-struct EventHandlers {
-    pub event: Event,
-    pub next_id: usize,
-    handlers: HashMap<usize, EventHandler>,
-}
-
-#[cfg(feature = "events")]
-impl EventHandlers {
-    fn new(event: Event) -> EventHandlers {
-        EventHandlers {
-            event,
-            next_id: 0,
-            handlers: HashMap::new(),
-        }
-    }
-
-    pub fn insert(&mut self, handler: HandlerFn) -> Option<EventHandler> {
-        let id = self.next_id;
-        self.next_id += 1;
-
-        let event_handler = EventHandler {
-            id,
-            event: self.event,
-            handler,
-        };
-
-        self.handlers.insert(id, event_handler);
-
-        Some(event_handler)
-    }
-}
+static CONNECTED: AtomicBool = AtomicBool::new(false);
 
 ///
 /// G29
@@ -249,18 +134,17 @@ impl EventHandlers {
 pub struct G29 {
     options: Options,
     prepend_write: bool,
-    inner: Arc<RwLock<InnerG29>>,
     calibrated: bool,
-    connected: bool,
+    inner: Arc<RwLock<InnerG29>>,
 }
 
 #[derive(Debug)]
 struct InnerG29 {
     data: Arc<RwLock<Frame>>,
-    wheel: Mutex<hidapi::HidDevice>,
     reader_handle: Option<thread::JoinHandle<()>>,
-    #[cfg(feature = "events")]
-    event_handlers: RwLock<HashMap<Event, EventHandlers>>,
+
+    event_handlers: HashMap<Event, RwLock<EventHandlers>>,
+    wheel: Option<Mutex<hidapi::HidDevice>>,
 }
 
 ///
@@ -318,15 +202,6 @@ fn get_wheel_info(api: &HidApi) -> DeviceInfo {
         .clone()
 }
 
-#[cfg(feature = "events")]
-fn different_indices(data1: &Frame, data2: &Frame) -> Vec<usize> {
-    data1
-        .iter()
-        .enumerate()
-        .filter_map(|(i, &x)| if x != data2[i] { Some(i) } else { None })
-        .collect()
-}
-
 impl G29 {
     ///
     /// Connect to the G29 Racing Wheel
@@ -360,15 +235,15 @@ impl G29 {
             options,
             prepend_write,
             calibrated: false,
-            connected: true,
             inner: Arc::new(RwLock::new(InnerG29 {
-                wheel: Mutex::new(wheel),
+                wheel: Some(Mutex::new(wheel)),
                 data: Arc::new(RwLock::new([0; FRAME_SIZE])),
                 reader_handle: None,
-                #[cfg(feature = "events")]
-                event_handlers: RwLock::new(HashMap::new()),
+
+                event_handlers: HashMap::new(),
             })),
         };
+        CONNECTED.store(true, std::sync::atomic::Ordering::Release);
 
         g29.initialize();
 
@@ -380,6 +255,8 @@ impl G29 {
             .read()
             .unwrap()
             .wheel
+            .as_ref()
+            .unwrap()
             .lock()
             .unwrap()
             .set_blocking_mode(false)
@@ -391,6 +268,8 @@ impl G29 {
             .read()
             .unwrap()
             .wheel
+            .as_ref()
+            .unwrap()
             .lock()
             .unwrap()
             .read(&mut data)
@@ -431,7 +310,15 @@ impl G29 {
                 .open_device(&HidApi::new().unwrap())
                 .unwrap();
 
-            *self.inner.read().unwrap().wheel.lock().unwrap() = new_wheel;
+            *self
+                .inner
+                .read()
+                .unwrap()
+                .wheel
+                .as_ref()
+                .unwrap()
+                .lock()
+                .unwrap() = new_wheel;
 
             self.initialize();
             return;
@@ -444,54 +331,53 @@ impl G29 {
             println!("listen -> Ready to listen for wheel events.");
         }
 
-        // let options = self.options.clone();
-        // let prepend_write = self.prepend_write;
-        // let calibrated = self.calibrated;
         // use thread to listen for wheel events and trigger events
         let self_1 = self.clone();
         let local_self = self.inner.clone();
-        let thread_handle = thread::spawn(move || loop {
-            let mut new_data = [0u8; FRAME_SIZE];
-            let size_read = local_self
-                .read()
-                .unwrap()
-                .wheel
-                .lock()
-                .unwrap()
-                .read(&mut new_data)
-                .expect("listen -> Error reading from device.");
-            if size_read == FRAME_SIZE {
-                let local_self_write = local_self.read().unwrap();
-                let mut prev_data = local_self_write.data.write().unwrap();
+        let thread_handle = thread::spawn(move || {
+            while CONNECTED.load(std::sync::atomic::Ordering::Relaxed) {
+                let mut new_data = [0u8; FRAME_SIZE];
+                let size_read = local_self
+                    .read()
+                    .unwrap()
+                    .wheel
+                    .as_ref()
+                    .unwrap()
+                    .lock()
+                    .unwrap()
+                    .read(&mut new_data)
+                    .expect("listen -> Error reading from device.");
+                if size_read == FRAME_SIZE {
+                    let local_self_write = local_self.read().unwrap();
+                    let mut prev_data = local_self_write.data.write().unwrap();
 
-                if new_data == *prev_data {
-                    continue;
+                    if new_data == *prev_data {
+                        continue;
+                    }
+
+                    {
+                        let events = events::events(&prev_data, &new_data);
+
+                        events.par_iter().for_each(|event| {
+                            local_self_write
+                                .event_handlers
+                                .get(event)
+                                .par_iter()
+                                .for_each(|event_handlers| {
+                                    event_handlers.read().unwrap().handlers.par_iter().for_each(
+                                        |(_, event_handler)| {
+                                            let mut self_1 = self_1.clone();
+                                            let ev_clone = *event_handler; // Clone the event handler
+                                            thread::spawn(move || {
+                                                (ev_clone.handler)(&mut self_1);
+                                            });
+                                        },
+                                    );
+                                })
+                        });
+                    }
+                    *prev_data = new_data;
                 }
-
-                #[cfg(feature = "events")]
-                {
-                    let events = local_self_write.events(&prev_data, &new_data);
-
-                    events.par_iter().for_each(|event| {
-                        local_self_write
-                            .event_handlers
-                            .read()
-                            .unwrap()
-                            .get(event)
-                            .map(|event_handlers| {
-                                event_handlers.handlers.values().for_each(|event_handler| {
-                                    let ev_clone = *event_handler;
-                                    let mut self_1 = self_1.clone();
-                                    thread::spawn(move || {
-                                        (ev_clone.handler)(&mut self_1);
-                                    });
-                                });
-
-                                Some(())
-                            });
-                    });
-                }
-                *prev_data = new_data;
             }
         });
 
@@ -601,6 +487,8 @@ impl G29 {
             .read()
             .unwrap()
             .wheel
+            .as_ref()
+            .expect("relay_os -> Wheel not found")
             .lock()
             .unwrap()
             .write(if self.prepend_write { &new_data } else { &data })
@@ -663,7 +551,7 @@ impl G29 {
     ///     sleep(Duration::from_secs(1));
     ///   }
     /// ````
-    pub fn set_leds(&mut self, leds: Led) {
+    pub fn set_leds(&self, leds: Led) {
         /*
             Set the LED lights on the G29.
         */
@@ -707,22 +595,25 @@ impl G29 {
     }
 
     /// Get the throttle value.
-    /// 0 is
+    ///  255 is depressed, 0 is fully pressed
     pub fn throttle(&self) -> u8 {
         state::throttle(&self.inner.read().unwrap().data.read().unwrap())
     }
 
     /// Get the brake value.
+    ///  255 is depressed, 0 is fully pressed
     pub fn brake(&self) -> u8 {
         state::brake(&self.inner.read().unwrap().data.read().unwrap())
     }
 
     /// Get the steering value.
+    /// 255 is fully right, 0 is fully left
     pub fn steering(&self) -> u8 {
         state::steering(&self.inner.read().unwrap().data.read().unwrap())
     }
 
     /// Get the fine steering value.
+    /// 255 is fully right, 0 is fully left
     pub fn steering_fine(&self) -> u8 {
         state::steering_fine(&self.inner.read().unwrap().data.read().unwrap())
     }
@@ -842,7 +733,8 @@ impl G29 {
         state::playstation_button(&self.inner.read().unwrap().data.read().unwrap())
     }
 
-    /// Returns the value of the clutch pedal. (0 - 255)
+    /// Returns the value of the clutch pedal.
+    /// 255 is depressed, 0 is fully pressed
     pub fn clutch(&self) -> u8 {
         state::clutch(&self.inner.read().unwrap().data.read().unwrap())
     }
@@ -880,443 +772,52 @@ impl G29 {
     ///   g29.disconnect();
     /// ```
     pub fn disconnect(&mut self) {
+        if !self.connected() {
+            return;
+        }
+
         self.force_off(0xf3);
         self.set_leds(Led::None);
         self.force_friction(0, 0);
         self.options.auto_center = [0x00, 0x00];
         self.set_auto_center();
-        let mut inner = self.inner.write().unwrap();
 
-        inner
-            .reader_handle
-            .take()
-            .and_then(|handle| handle.join().ok());
+        // set connected to false
 
-        inner.reader_handle = None;
-        self.connected = false;
+        CONNECTED.store(false, std::sync::atomic::Ordering::Release);
+        self.inner.write().unwrap().wheel = None;
+        // join all threads
+        if let Some(handle) = self.inner.write().unwrap().reader_handle.take() {
+            handle.join().unwrap();
+        }
     }
 
     pub fn connected(&self) -> bool {
-        self.connected
+        CONNECTED.load(std::sync::atomic::Ordering::Relaxed)
+            && self.inner.read().unwrap().wheel.is_some()
     }
 
-    #[cfg(feature = "events")]
     pub fn register_event_handler(&self, event: Event, handler: HandlerFn) -> Option<EventHandler> {
         self.inner
             .write()
             .unwrap()
             .event_handlers
+            .entry(event)
+            .or_insert_with(|| RwLock::new(EventHandlers::new(event)))
             .write()
             .unwrap()
-            .entry(event)
-            .or_insert_with(|| EventHandlers::new(event))
             .insert(handler)
     }
 
-    #[cfg(feature = "events")]
     pub fn unregister_event_handler(&mut self, event_handler: EventHandler) {
         self.inner
             .write()
             .unwrap()
             .event_handlers
-            .write()
-            .unwrap()
             .get_mut(&event_handler.event)
             .map(|handlers| {
-                handlers.handlers.remove(&event_handler.id);
+                handlers.write().unwrap().handlers.remove(&event_handler.id);
                 Some(handlers)
             });
-    }
-}
-
-impl Drop for G29 {
-    fn drop(&mut self) {
-        self.disconnect();
-    }
-}
-
-#[cfg(feature = "events")]
-impl InnerG29 {
-    fn events(&self, prev_data: &Frame, new_data: &Frame) -> Vec<Event> {
-        let different_indices = different_indices(prev_data, new_data);
-
-        if different_indices.is_empty() {
-            return vec![];
-        }
-
-        let events_to_trigger = different_indices.par_iter().flat_map(|index| {
-            let mut events_to_trigger = vec![];
-            match index {
-                0 => {
-                    events_to_trigger.extend(self.dpad_events(prev_data, new_data));
-                    events_to_trigger.extend(self.shape_button_events(prev_data, new_data));
-                }
-                1 => events_to_trigger.extend(self.data1_button_events(prev_data, new_data)),
-                2 => {
-                    events_to_trigger.extend(self.gear_selector_events(prev_data, new_data));
-                    events_to_trigger.extend(self.plus_button_events(prev_data, new_data));
-                }
-                3 => events_to_trigger.extend(self.data3_button_events(prev_data, new_data)),
-                4 | 5 => events_to_trigger.extend(self.steering_events(prev_data, new_data)),
-                6 => events_to_trigger.extend(self.throttle_event(prev_data, new_data)),
-                7 => events_to_trigger.extend(self.brake_event(prev_data, new_data)),
-                8 => events_to_trigger.extend(self.clutch_event(prev_data, new_data)),
-                9 => events_to_trigger.extend(self.shifter_x_event(prev_data, new_data)),
-                10 => events_to_trigger.extend(self.shifter_y_event(prev_data, new_data)),
-                11 => events_to_trigger.extend(self.shifter_events(prev_data, new_data)),
-                _ => {}
-            };
-            events_to_trigger
-        });
-
-        events_to_trigger.collect()
-    }
-
-    fn dpad_events(&self, prev_data: &Frame, new_data: &Frame) -> Vec<Event> {
-        let prev_dpad = state::dpad(prev_data);
-        let new_dpad = state::dpad(new_data);
-        if prev_dpad == new_dpad {
-            return vec![];
-        }
-
-        let mut events = vec![];
-
-        // which dpad is pressed
-        match new_dpad {
-            DpadPosition::Up => events.push(Event::DpadUpPressed),
-            DpadPosition::TopRight => events.push(Event::DpadTopRightPressed),
-            DpadPosition::Right => events.push(Event::DpadRightPressed),
-            DpadPosition::BottomRight => events.push(Event::DpadBottomRightPressed),
-            DpadPosition::Down => events.push(Event::DpadBottomPressed),
-            DpadPosition::BottomLeft => events.push(Event::DpadBottomLeftPressed),
-            DpadPosition::Left => events.push(Event::DpadLeftPressed),
-            DpadPosition::TopLeft => events.push(Event::DpadTopLeftPressed),
-            _ => {}
-        };
-
-        // which dpad is released
-        match prev_dpad {
-            DpadPosition::Up => events.push(Event::DpadUpReleased),
-            DpadPosition::TopRight => events.push(Event::DpadTopRightReleased),
-            DpadPosition::Right => events.push(Event::DpadRightReleased),
-            DpadPosition::BottomRight => events.push(Event::DpadBottomRightReleased),
-            DpadPosition::Down => events.push(Event::DpadBottomReleased),
-            DpadPosition::BottomLeft => events.push(Event::DpadBottomLeftReleased),
-            DpadPosition::Left => events.push(Event::DpadLeftReleased),
-            DpadPosition::TopLeft => events.push(Event::DpadTopLeftReleased),
-            _ => {}
-        };
-
-        events
-    }
-
-    fn shape_button_events(&self, prev_data: &Frame, new_data: &Frame) -> Vec<Event> {
-        let mut events = vec![];
-
-        let prev_x_button = state::x_button(prev_data);
-        let new_x_button = state::x_button(new_data);
-        if prev_x_button != new_x_button {
-            if new_x_button {
-                events.push(Event::XButtonPressed);
-            } else {
-                events.push(Event::XButtonReleased);
-            }
-        }
-
-        let prev_square_button = state::square_button(prev_data);
-        let new_square_button = state::square_button(new_data);
-        if prev_square_button != new_square_button {
-            if new_square_button {
-                events.push(Event::SquareButtonPressed);
-            } else {
-                events.push(Event::SquareButtonReleased);
-            }
-        }
-
-        let prev_circle_button = state::circle_button(prev_data);
-        let new_circle_button = state::circle_button(new_data);
-        if prev_circle_button != new_circle_button {
-            if new_circle_button {
-                events.push(Event::CircleButtonPressed);
-            } else {
-                events.push(Event::CircleButtonReleased);
-            }
-        }
-
-        let prev_triangle_button = state::triangle_button(prev_data);
-        let new_triangle_button = state::triangle_button(new_data);
-        if prev_triangle_button != new_triangle_button {
-            if new_triangle_button {
-                events.push(Event::TriangleButtonPressed);
-            } else {
-                events.push(Event::TriangleButtonReleased);
-            }
-        }
-
-        events
-    }
-
-    fn data1_button_events(&self, prev_data: &Frame, new_data: &Frame) -> Vec<Event> {
-        let mut events = vec![];
-
-        let prev_right_shifter = state::right_shifter(prev_data);
-        let new_right_shifter = state::right_shifter(new_data);
-        if prev_right_shifter != new_right_shifter {
-            if new_right_shifter {
-                events.push(Event::RightShifterPressed);
-            } else {
-                events.push(Event::RightShifterReleased);
-            }
-        }
-
-        let prev_left_shifter = state::left_shifter(prev_data);
-        let new_left_shifter = state::left_shifter(new_data);
-        if prev_left_shifter != new_left_shifter {
-            if new_left_shifter {
-                events.push(Event::LeftShifterPressed);
-            } else {
-                events.push(Event::LeftShifterReleased);
-            }
-        }
-
-        let prev_r2_button = state::r2_button(prev_data);
-        let new_r2_button = state::r2_button(new_data);
-        if prev_r2_button != new_r2_button {
-            if new_r2_button {
-                events.push(Event::R2ButtonPressed);
-            } else {
-                events.push(Event::R2ButtonReleased);
-            }
-        }
-
-        let prev_l2_button = state::l2_button(prev_data);
-        let new_l2_button = state::l2_button(new_data);
-        if prev_l2_button != new_l2_button {
-            if new_l2_button {
-                events.push(Event::L2ButtonPressed);
-            } else {
-                events.push(Event::L2ButtonReleased);
-            }
-        }
-
-        let prev_share_button = state::share_button(prev_data);
-        let new_share_button = state::share_button(new_data);
-        if prev_share_button != new_share_button {
-            if new_share_button {
-                events.push(Event::ShareButtonPressed);
-            } else {
-                events.push(Event::ShareButtonReleased);
-            }
-        }
-
-        let prev_option_button = state::option_button(prev_data);
-        let new_option_button = state::option_button(new_data);
-        if prev_option_button != new_option_button {
-            if new_option_button {
-                events.push(Event::OptionButtonPressed);
-            } else {
-                events.push(Event::OptionButtonReleased);
-            }
-        }
-
-        let prev_r3_button = state::r3_button(prev_data);
-        let new_r3_button = state::r3_button(new_data);
-        if prev_r3_button != new_r3_button {
-            if new_r3_button {
-                events.push(Event::R3ButtonPressed);
-            } else {
-                events.push(Event::R3ButtonReleased);
-            }
-        }
-
-        let prev_l3_button = state::l3_button(prev_data);
-        let new_l3_button = state::l3_button(new_data);
-        if prev_l3_button != new_l3_button {
-            if new_l3_button {
-                events.push(Event::L3ButtonPressed);
-            } else {
-                events.push(Event::L3ButtonReleased);
-            }
-        }
-
-        events
-    }
-
-    fn gear_selector_events(&self, prev_data: &Frame, new_data: &Frame) -> Vec<Event> {
-        let prev_gear_selector = state::gear_selector(prev_data);
-        let new_gear_selector = state::gear_selector(new_data);
-
-        if prev_gear_selector == new_gear_selector {
-            return vec![];
-        }
-
-        let mut events = vec![];
-
-        match new_gear_selector {
-            GearSelector::Neutral => events.push(Event::GearChanged),
-            GearSelector::First => events.push(Event::GearChanged),
-            GearSelector::Second => events.push(Event::GearChanged),
-            GearSelector::Third => events.push(Event::GearChanged),
-            GearSelector::Fourth => events.push(Event::GearChanged),
-            GearSelector::Fifth => events.push(Event::GearChanged),
-            GearSelector::Sixth => events.push(Event::GearChanged),
-            GearSelector::Reverse => events.push(Event::GearChanged),
-        }
-
-        events
-    }
-
-    fn plus_button_events(&self, prev_data: &Frame, new_data: &Frame) -> Vec<Event> {
-        let prev_plus_button = state::plus_button(prev_data);
-        let new_plus_button = state::plus_button(new_data);
-        if prev_plus_button == new_plus_button {
-            vec![]
-        } else if new_plus_button {
-            vec![Event::PlusButtonPressed]
-        } else {
-            vec![Event::PlusButtonReleased]
-        }
-    }
-
-    fn data3_button_events(&self, prev_data: &Frame, new_data: &Frame) -> Vec<Event> {
-        /*
-           minus_button
-           spinner_right
-           spinner_left
-           spinner_button
-           playstation_button
-        */
-
-        let mut events = vec![];
-
-        let prev_minus_button = state::minus_button(prev_data);
-        let new_minus_button = state::minus_button(new_data);
-        if prev_minus_button != new_minus_button {
-            if new_minus_button {
-                events.push(Event::MinusButtonPressed);
-            } else {
-                events.push(Event::MinusButtonReleased);
-            }
-        }
-
-        let prev_spinner_right = state::spinner_right(prev_data);
-        let new_spinner_right = state::spinner_right(new_data);
-        if prev_spinner_right != new_spinner_right && new_spinner_right {
-            events.push(Event::SpinnerRight);
-        }
-
-        let prev_spinner_left = state::spinner_left(prev_data);
-        let new_spinner_left = state::spinner_left(new_data);
-        if prev_spinner_left != new_spinner_left && new_spinner_left {
-            events.push(Event::SpinnerLeft);
-        }
-
-        let prev_spinner_button = state::spinner_button(prev_data);
-        let new_spinner_button = state::spinner_button(new_data);
-        if prev_spinner_button != new_spinner_button {
-            if new_spinner_button {
-                events.push(Event::SpinnerButtonPressed);
-            } else {
-                events.push(Event::SpinnerButtonReleased);
-            }
-        }
-
-        let prev_playstation_button = state::playstation_button(prev_data);
-        let new_playstation_button = state::playstation_button(new_data);
-        if prev_playstation_button != new_playstation_button {
-            if new_playstation_button {
-                events.push(Event::PlaystationButtonPressed);
-            } else {
-                events.push(Event::PlaystationButtonReleased);
-            }
-        }
-
-        events
-    }
-
-    fn steering_events(&self, prev_data: &Frame, new_data: &Frame) -> Vec<Event> {
-        let mut events = vec![];
-
-        let prev_steering = state::steering(prev_data);
-        let new_steering = state::steering(new_data);
-        if prev_steering != new_steering {
-            events.push(Event::Steering);
-        }
-
-        let prev_steering_fine = state::steering_fine(prev_data);
-        let new_steering_fine = state::steering_fine(new_data);
-        if prev_steering_fine != new_steering_fine {
-            events.push(Event::SteeringFine);
-        }
-
-        events
-    }
-
-    fn throttle_event(&self, prev_data: &Frame, new_data: &Frame) -> Vec<Event> {
-        let prev_throttle = state::throttle(prev_data);
-        let new_throttle = state::throttle(new_data);
-
-        if prev_throttle == new_throttle {
-            vec![]
-        } else {
-            vec![Event::Throttle]
-        }
-    }
-
-    fn brake_event(&self, prev_data: &Frame, new_data: &Frame) -> Vec<Event> {
-        let prev_brake = state::brake(prev_data);
-        let new_brake = state::brake(new_data);
-
-        if prev_brake == new_brake {
-            vec![]
-        } else {
-            vec![Event::Brake]
-        }
-    }
-
-    fn clutch_event(&self, prev_data: &Frame, new_data: &Frame) -> Vec<Event> {
-        let prev_clutch = state::clutch(prev_data);
-        let new_clutch = state::clutch(new_data);
-
-        if prev_clutch == new_clutch {
-            vec![]
-        } else {
-            vec![Event::Clutch]
-        }
-    }
-
-    fn shifter_x_event(&self, prev_data: &Frame, new_data: &Frame) -> Vec<Event> {
-        let prev_shifter_x = state::shifter_x(prev_data);
-        let new_shifter_x = state::shifter_x(new_data);
-
-        if prev_shifter_x == new_shifter_x {
-            vec![]
-        } else {
-            vec![Event::ShifterX]
-        }
-    }
-
-    fn shifter_y_event(&self, prev_data: &Frame, new_data: &Frame) -> Vec<Event> {
-        let prev_shifter_y = state::shifter_y(prev_data);
-        let new_shifter_y = state::shifter_y(new_data);
-
-        if prev_shifter_y == new_shifter_y {
-            vec![]
-        } else {
-            vec![Event::ShifterY]
-        }
-    }
-
-    fn shifter_events(&self, prev_data: &Frame, new_data: &Frame) -> Vec<Event> {
-        let prev_shifter_pressed = state::shifter_pressed(prev_data);
-        let new_shifter_pressed = state::shifter_pressed(new_data);
-
-        if prev_shifter_pressed == new_shifter_pressed {
-            vec![]
-        } else if new_shifter_pressed {
-            vec![Event::ShifterPressed]
-        } else {
-            vec![Event::ShifterReleased]
-        }
     }
 }
